+++ data/types/food/protein_test.go
@@ -17,7 +17,7 @@ import (
 
 func NewProtein() *food.Protein {
 	datum := food.NewProtein()
-	datum.Total = pointer.FromInt(test.RandomIntFromRange(food.ProteinTotalGramsMinimum, food.ProteinTotalGramsMaximum))
+	datum.Total = pointer.FromFloat64(test.RandomFloat64FromRange(food.ProteinTotalGramsMinimum, food.ProteinTotalGramsMaximum))
 	datum.Units = pointer.FromString(test.RandomStringFromArray(food.ProteinUnits()))
 	return datum
 }
@@ -27,18 +27,18 @@ func CloneProtein(datum *food.Protein) *food.Protein {
 		return nil
 	}
 	clone := food.NewProtein()
-	clone.Total = test.CloneInt(datum.Total)
+	clone.Total = test.CloneFloat64(datum.Total)
 	clone.Units = test.CloneString(datum.Units)
 	return clone
 }
 
 var _ = Describe("Protein", func() {
 	It("ProteinTotalGramsMaximum is expected", func() {
-		Expect(food.ProteinTotalGramsMaximum).To(Equal(1000))
+		Expect(food.ProteinTotalGramsMaximum).To(Equal(1000.0))
 	})
 
 	It("ProteinTotalGramsMinimum is expected", func() {
-		Expect(food.ProteinTotalGramsMinimum).To(Equal(0))
+		Expect(food.ProteinTotalGramsMinimum).To(Equal(0.0))
 	})
 
 	It("ProteinUnitsGrams is expected", func() {
@@ -79,18 +79,18 @@ var _ = Describe("Protein", func() {
 					testErrors.WithPointerSource(structureValidator.ErrorValueNotExists(), "/total"),
 				),
 				Entry("total out of range (lower)",
-					func(datum *food.Protein) { datum.Total = pointer.FromInt(-1) },
-					testErrors.WithPointerSource(structureValidator.ErrorValueNotInRange(-1, 0, 1000), "/total"),
+					func(datum *food.Protein) { datum.Total = pointer.FromFloat64(-0.1) },
+					testErrors.WithPointerSource(structureValidator.ErrorValueNotInRange(-0.1, 0.0, 1000.0), "/total"),
 				),
 				Entry("total in range (lower)",
-					func(datum *food.Protein) { datum.Total = pointer.FromInt(0) },
+					func(datum *food.Protein) { datum.Total = pointer.FromFloat64(0.0) },
 				),
 				Entry("total in range (upper)",
-					func(datum *food.Protein) { datum.Total = pointer.FromInt(1000) },
+					func(datum *food.Protein) { datum.Total = pointer.FromFloat64(1000.0) },
 				),
 				Entry("total out of range (upper)",
-					func(datum *food.Protein) { datum.Total = pointer.FromInt(1001) },
-					testErrors.WithPointerSource(structureValidator.ErrorValueNotInRange(1001, 0, 1000), "/total"),
+					func(datum *food.Protein) { datum.Total = pointer.FromFloat64(1000.1) },
+					testErrors.WithPointerSource(structureValidator.ErrorValueNotInRange(1000.1, 0.0, 1000.0), "/total"),
 				),
 				Entry("units missing",
 					func(datum *food.Protein) { datum.Units = nil },